Refactor CategoryController and add list feature
The CategoryController class has been refactored to use a wildcard import for Spring's web.bind.annotation package. A new feature has also been added that allows fetching all categories from the database. This is done through a GET request at the root endpoint of the controller, which returns a list of Category objects wrapped in a ResponseEntity with an HTTP status code of OK.